src/main/java/com/github/brunomndantas/jscrapper/support/parser/single/text/reference/SingleReferenceBooleanTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.single.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.single.text.SingleTextParser;
import org.openqa.selenium.WebElement;

public class SingleReferenceBooleanTextParser extends SingleTextParser {

    public SingleReferenceBooleanTextParser() {
        super(null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Boolean.parseBoolean(element.getText());
    }

}