src/test/java/com/github/brunomndantas/jscrapper/support/parser/single/text/reference/SingleReferenceBooleanTextParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.single.text.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class SingleReferenceBooleanTextParserTest {

    @Test
    public void constructorTest() {
        SingleReferenceBooleanTextParser parser = new SingleReferenceBooleanTextParser();
        assertEquals(null, parser.getDefaultValue());
    }

    @Test
    public void parseElement() throws Exception {
        SingleReferenceBooleanTextParser parser = new SingleReferenceBooleanTextParser();

        assertEquals(true, parser.parseElement(new DummyElement(){
            @Override
            public String getText() {
                return "true";
            }
        }));
    }

}